fix: split function_3
