+++ packages/amplify-dotnet-function-template-provider/resources/lambda/Trigger/Function.csproj.ejs
@@ -5,15 +5,15 @@
     <AWSProjectType>Lambda</AWSProjectType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Amazon.Lambda.Core" Version="1.1.0" />
-    <PackageReference Include="Amazon.Lambda.Serialization.Json" Version="1.7.0" />
+    <PackageReference Include="Amazon.Lambda.Core" Version="2.1.0" />
+    <PackageReference Include="Amazon.Lambda.Serialization.Json" Version="2.1.0" />
     <% switch(props.triggerEventSourceMappings[0].functionTemplateType) {
          case 'dynamoDB':%>
     <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.3.101.80" />
-    <PackageReference Include="Amazon.Lambda.DynamoDBEvents" Version="1.1.0" />
+    <PackageReference Include="Amazon.Lambda.DynamoDBEvents" Version="2.1.1" />
     <%     break;
          case 'kinesis':%>
-    <PackageReference Include="Amazon.Lambda.KinesisEvents" Version="1.1.0" />
+    <PackageReference Include="Amazon.Lambda.KinesisEvents" Version="2.0.0" />
     <%     break;
          default:
            console.log(`Unknown template type '${props.triggerEventSourceMappings[0].functionTemplateType}'`);
